[react-frontend] make API call to localhost
Fetch all recipes when app loads (App component's componentDidMount);
store these in state as `allRecipes`.

When query form is acutally built in the future it'll simply filter this
list rather than make API calls of its own.

This won't scale, but I don't think it'll need to for our personal use,
so :shrug:
--
Also add some classes to various parts and a style rule for one of them
to begin making this look decent.